Refactor chatbot and review modules: consolidate functionality and improve API structure
